extract aop template code
